requirements.txt
openai>=1.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
python-dotenv>=0.19.0
sqlalchemy>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
ollama>=0.1.0
rich>=10.0.0
SQLAlchemy>=1.4.0 